# wayfarer-swagger-doc1.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "WayFarer",
		"description": "WayFarer is a public bus transportation booking server.    ",
		"contact": {
			"email": "kolaakindoju@gmail.com"
		},
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "https://wayfarer-1.herokuapp.com/"
		}
	],
...
	"paths": {
...
		"/api/v1/bookings": {
			"post": {
				"tags": [
					"Bookings"
				],
				"summary": "Book a trip",
				"description": "This can only be done by a logged in user.",
				"operationId": "bookTrip",
				"requestBody": {
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/Bookings-post-request"
							}
						},
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Bookings-post-request"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Bookings-post-response"
								}
							}
						}
					},
					"4XX": {
						"description": "Client errors",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error-response"
								}
							}
						}
					},
					"5XX": {
						"description": "Server errors",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error-response"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"get": {
				"tags": [
					"Bookings"
				],
				"summary": "Get bookings",
				"description": "This can only be done by a logged in user or admin.",
				"operationId": "getBookings",
				"responses": {
					"200": {
						"description": "Successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Bookings-get-response"
								}
							}
						}
					},
					"4XX": {
						"description": "Client errors",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error-response"
								}
							}
						}
					},
					"5XX": {
						"description": "Server errors",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error-response"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
			"Error-response": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"...": {
						"type": "object"
					}
				}
			},
...
			"Bookings-post-request": {
				"type": "object",
				"properties": {
					"trip_id": {
						"type": "integer"
					},
					"seat_number": {
						"type": "integer"
					}
				},
				"required": [
					"trip_id"
				]
			},
			"Bookings-post-response": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string"
					},
					"data": {
						"type": "object",
						"properties": {
							"id": {
								"type": "integer"
							},
							"booking_id": {
								"type": "integer"
							},
							"user_id": {
								"type": "integer"
							},
							"trip_id": {
								"type": "integer"
							},
							"bus_id": {
								"type": "integer"
							},
							"trip_date": {
								"type": "string",
								"format": "date"
							},
							"seat_number": {
								"type": "integer"
							},
							"first_name": {
								"type": "string"
							},
							"last_name": {
								"type": "string"
							},
							"email": {
								"type": "string",
								"format": "email"
							}
						}
					}
				}
			},
			"Bookings-get-response": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"id": {
									"type": "integer"
								},
								"user_id": {
									"type": "integer"
								},
								"trip_id": {
									"type": "integer"
								},
								"bus_id": {
									"type": "integer"
								},
								"trip_date": {
									"type": "string",
									"format": "date"
								},
								"seat_number": {
									"type": "integer"
								},
								"first_name": {
									"type": "string"
								},
								"last_name": {
									"type": "string"
								},
								"email": {
									"type": "string",
									"format": "email"
								}
							}
						}
					}
				}
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}